requirements.txt
torch>=2.0
torchvision>=0.15
numpy
scipy
scikit-learn
pandas
matplotlib
tqdm
kornia
Pillow
